Replace vue-analytics with manual tracking
Turns out `window` in content scripts is not the same `window`
as on a web page. They share the same `document`, but not `window`.
Can't see any other way to solve this than to either load ga
someway differently, or load it manually and send postMessage
to inform about navigation.
This solution is the manual approach, it will only add the script
and not handle navigation or errors.