New "API" for library
Looks like /lib-ajax is no more (Audible removed it).
Also had to add a check for author where a book could be missing
a title. I got this error once in my library, but I haven't seen it
before, so this could mean something else in the markup has changed.
No time to look into it now, it should at least not break now.