Add filter for hiding books based on already own book series index
If a book is already in users library with the same series book index
for the same series then this books should be possible to filter out.
Ugh, this sounds complicated, but it's not that hard.
Imagine you have book 1,2,3 and 4 of a series, then book with index 2
is re-released (better version or something), then you now have the
option to hide this book, because you have already read the original.